quilting_backend/requirements.txt
django
djangorestframework
djangorestframework-jwt
djangorestframework-camel-case
django-cors-headers
Pillow
